+++ src/common/errorMessage/index.ts
@@ -1,0 +1,1 @@
+export {ErrorMessage} from './ErrorMessage'
