+++ src/utils/index.ts
@@ -1,0 +1,2 @@
+export {getLowerCaseString} from './getLowerCaseString';
+export {getFormatedDate, getDateToTimeStamp} from './getFormatedDate';
