+++ src/widgets/notes/components/formWithNotes/index.ts
@@ -1,0 +1,1 @@
+export {FormWithNotes} from './FormWithNotes';
