+++ src/widgets/notes/enums/index.ts
@@ -1,0 +1,1 @@
+export {LocalStorageTitles} from './enum';
