+++ src/widgets/notes/index.ts
@@ -1,10 +1,1 @@
-export * from './typings/';
-export * from './enums/enum';
-export * from './service/getNotes';
-export * from '../../services/localStorage';
-export * from '../../utils/getLowerCaseString';
-export {ErrorMessage} from '../../common/errorMessage/ErrorMessage';
-export {FormWithNotes} from './components/formWithNotes/FormWithNotes';
-export {
-    AccordionWithNotes
-} from './components/accordionWithButtonForm/AccordionWithNotes';
+export {FormWithNotes, AccordionWithNotes} from './components';
