+++ src/widgets/notes/service/index.ts
@@ -1,0 +1,1 @@
+export {getNotes} from './getNotes';
